requirements.txt
aniso8601==7.0.0
apt-xapian-index==0.47
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.2.5
blinker==1.4
Brlapi==0.6.7
certifi==2018.8.24
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
cryptography==2.3
cupshelpers==1.0
dataclasses==0.6
defer==1.0.6
distro==1.3.0
distro-info===0.21ubuntu2
dj-database-url==0.5.0
Django==2.2.4
django-contact-form==1.7
django-environ==0.4.5
django-extensions==2.2.1
django-feedparser==0.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-jet==1.0.8
djangorestframework==3.10.2
djongo==1.2.33
entrypoints==0.3
feedparser==5.2.1
google-api-python-client==1.4.1
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
httplib2==0.11.3
idna==2.6
isort==4.3.21
keyring==17.1.1
keyrings.alt==3.1.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
louis==3.8.0
macaroonbakery==1.2.1
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.0
mccabe==0.6.1
netifaces==0.10.4
oauth2client==4.1.3
oauthlib==2.1.0
olefile==0.46
pexpect==4.6.0
Pillow==5.4.1
promise==2.2.1
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.32.0
PyJWT==1.7.0
pylint==2.3.1
pymacaroons==0.13.0
pymongo==3.9.0
PyNaCl==1.3.0
PyQt5==5.12.1
pyRFC3339==1.1
python-apt==1.8.4
python-dateutil==2.7.3
python-debian==0.1.34
pytz==2018.9
pyxdg==0.25
PyYAML==3.13
reportlab==3.5.18
requests==2.21.0
requests-unixsocket==0.1.5
rsa==4.0
Rx==1.6.1
SecretStorage==2.3.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.2.4
system-service==0.3
systemd-python==234
typed-ast==1.4.0
ubuntu-advantage-tools==19.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.24.1
usb-creator==0.3.3
wadllib==1.3.3
whitenoise==4.1.3
wrapt==1.11.2
xkit==0.0.0
